Do refactor to remove hedemail, rename schematools and hed_cache.  Put tests in separate folders.
