--- hedemail/deploy/hedemailer.wsgi
@@ -1,6 +1,0 @@
-#!/usr/bin/python
-import sys
-activate_this = '/var/www/gollum/env/bin/activate_this.py'
-exec(open(activate_this).read(), dict(__file__=activate_this))
-sys.path.insert(0, "/var/www/gollum/hedemail")
-from hed.emailer.create_app import app as application
