--- hedemail/tests/data/emails.txt
@@ -1,1 +1,0 @@
-abc@gmail.com
