--- hedemail/tests/data/hed_payload.json
@@ -1,131 +1,0 @@
-{
-  "pages": [
-    {
-      "page_name": "HED-Schema",
-      "title": "HED Schema",
-      "summary": null,
-      "action": "edited",
-      "sha": "c47813b3518967c2ba59fb1514465deb14495ce2",
-      "html_url": "https://github.com/jcockfie/HED-schema/wiki/HED-Schema"
-    }
-  ],
-  "repository": {
-    "id": 130238929,
-    "node_id": "MDEwOlJlcG9zaXRvcnkxMzAyMzg5Mjk=",
-    "name": "HED-schema",
-    "full_name": "jcockfie/HED-schema",
-    "owner": {
-      "login": "jcockfie",
-      "id": 1193807,
-      "node_id": "MDQ6VXNlcjExOTM4MDc=",
-      "avatar_url": "https://avatars2.githubusercontent.com/u/1193807?v=4",
-      "gravatar_id": "",
-      "url": "https://api.github.com/users/jcockfie",
-      "html_url": "https://github.com/jcockfie",
-      "followers_url": "https://api.github.com/users/jcockfie/followers",
-      "following_url": "https://api.github.com/users/jcockfie/following{/other_user}",
-      "gists_url": "https://api.github.com/users/jcockfie/gists{/gist_id}",
-      "starred_url": "https://api.github.com/users/jcockfie/starred{/owner}{/repo}",
-      "subscriptions_url": "https://api.github.com/users/jcockfie/subscriptions",
-      "organizations_url": "https://api.github.com/users/jcockfie/orgs",
-      "repos_url": "https://api.github.com/users/jcockfie/repos",
-      "events_url": "https://api.github.com/users/jcockfie/events{/privacy}",
-      "received_events_url": "https://api.github.com/users/jcockfie/received_events",
-      "type": "User",
-      "site_admin": false
-    },
-    "private": false,
-    "html_url": "https://github.com/jcockfie/HED-schema",
-    "description": "Hierarchical Event Descriptor (HED) Tags and related tools.",
-    "fork": true,
-    "url": "https://api.github.com/repos/jcockfie/HED-schema",
-    "forks_url": "https://api.github.com/repos/jcockfie/HED-schema/forks",
-    "keys_url": "https://api.github.com/repos/jcockfie/HED-schema/keys{/key_id}",
-    "collaborators_url": "https://api.github.com/repos/jcockfie/HED-schema/collaborators{/collaborator}",
-    "teams_url": "https://api.github.com/repos/jcockfie/HED-schema/teams",
-    "hooks_url": "https://api.github.com/repos/jcockfie/HED-schema/hooks",
-    "issue_events_url": "https://api.github.com/repos/jcockfie/HED-schema/issues/events{/number}",
-    "events_url": "https://api.github.com/repos/jcockfie/HED-schema/events",
-    "assignees_url": "https://api.github.com/repos/jcockfie/HED-schema/assignees{/user}",
-    "branches_url": "https://api.github.com/repos/jcockfie/HED-schema/branches{/branch}",
-    "tags_url": "https://api.github.com/repos/jcockfie/HED-schema/tags",
-    "blobs_url": "https://api.github.com/repos/jcockfie/HED-schema/git/blobs{/sha}",
-    "git_tags_url": "https://api.github.com/repos/jcockfie/HED-schema/git/tags{/sha}",
-    "git_refs_url": "https://api.github.com/repos/jcockfie/HED-schema/git/refs{/sha}",
-    "trees_url": "https://api.github.com/repos/jcockfie/HED-schema/git/trees{/sha}",
-    "statuses_url": "https://api.github.com/repos/jcockfie/HED-schema/statuses/{sha}",
-    "languages_url": "https://api.github.com/repos/jcockfie/HED-schema/languages",
-    "stargazers_url": "https://api.github.com/repos/jcockfie/HED-schema/stargazers",
-    "contributors_url": "https://api.github.com/repos/jcockfie/HED-schema/contributors",
-    "subscribers_url": "https://api.github.com/repos/jcockfie/HED-schema/subscribers",
-    "subscription_url": "https://api.github.com/repos/jcockfie/HED-schema/subscription",
-    "commits_url": "https://api.github.com/repos/jcockfie/HED-schema/commits{/sha}",
-    "git_commits_url": "https://api.github.com/repos/jcockfie/HED-schema/git/commits{/sha}",
-    "comments_url": "https://api.github.com/repos/jcockfie/HED-schema/comments{/number}",
-    "issue_comment_url": "https://api.github.com/repos/jcockfie/HED-schema/issues/comments{/number}",
-    "contents_url": "https://api.github.com/repos/jcockfie/HED-schema/contents/{+path}",
-    "compare_url": "https://api.github.com/repos/jcockfie/HED-schema/compare/{base}...{head}",
-    "merges_url": "https://api.github.com/repos/jcockfie/HED-schema/merges",
-    "archive_url": "https://api.github.com/repos/jcockfie/HED-schema/{archive_format}{/ref}",
-    "downloads_url": "https://api.github.com/repos/jcockfie/HED-schema/downloads",
-    "issues_url": "https://api.github.com/repos/jcockfie/HED-schema/issues{/number}",
-    "pulls_url": "https://api.github.com/repos/jcockfie/HED-schema/pulls{/number}",
-    "milestones_url": "https://api.github.com/repos/jcockfie/HED-schema/milestones{/number}",
-    "notifications_url": "https://api.github.com/repos/jcockfie/HED-schema/notifications{?since,all,participating}",
-    "labels_url": "https://api.github.com/repos/jcockfie/HED-schema/labels{/name}",
-    "releases_url": "https://api.github.com/repos/jcockfie/HED-schema/releases{/id}",
-    "deployments_url": "https://api.github.com/repos/jcockfie/HED-schema/deployments",
-    "created_at": "2018-04-19T15:48:51Z",
-    "updated_at": "2018-04-19T15:48:53Z",
-    "pushed_at": "2018-04-19T15:56:58Z",
-    "git_url": "git://github.com/jcockfie/HED-schema.git",
-    "ssh_url": "git@github.com:jcockfie/HED-schema.git",
-    "clone_url": "https://github.com/jcockfie/HED-schema.git",
-    "svn_url": "https://github.com/jcockfie/HED-schema",
-    "homepage": "",
-    "size": 1221,
-    "stargazers_count": 0,
-    "watchers_count": 0,
-    "language": null,
-    "has_issues": false,
-    "has_projects": true,
-    "has_downloads": true,
-    "has_wiki": true,
-    "has_pages": true,
-    "forks_count": 0,
-    "mirror_url": null,
-    "archived": false,
-    "open_issues_count": 0,
-    "license": {
-      "key": "mit",
-      "name": "MIT License",
-      "spdx_id": "MIT",
-      "url": "https://api.github.com/licenses/mit",
-      "node_id": "MDc6TGljZW5zZTEz"
-    },
-    "forks": 0,
-    "open_issues": 0,
-    "watchers": 0,
-    "default_branch": "master"
-  },
-  "sender": {
-    "login": "jcockfie",
-    "id": 1193807,
-    "node_id": "MDQ6VXNlcjExOTM4MDc=",
-    "avatar_url": "https://avatars2.githubusercontent.com/u/1193807?v=4",
-    "gravatar_id": "",
-    "url": "https://api.github.com/users/jcockfie",
-    "html_url": "https://github.com/jcockfie",
-    "followers_url": "https://api.github.com/users/jcockfie/followers",
-    "following_url": "https://api.github.com/users/jcockfie/following{/other_user}",
-    "gists_url": "https://api.github.com/users/jcockfie/gists{/gist_id}",
-    "starred_url": "https://api.github.com/users/jcockfie/starred{/owner}{/repo}",
-    "subscriptions_url": "https://api.github.com/users/jcockfie/subscriptions",
-    "organizations_url": "https://api.github.com/users/jcockfie/orgs",
-    "repos_url": "https://api.github.com/users/jcockfie/repos",
-    "events_url": "https://api.github.com/users/jcockfie/events{/privacy}",
-    "received_events_url": "https://api.github.com/users/jcockfie/received_events",
-    "type": "User",
-    "site_admin": false
-  }
-}
