+++ hedtools/hed/schema/parsetag.py
