+++ hedtools/tests/util/test_file_util.py
@@ -6,7 +6,7 @@ class Test(unittest.TestCase):
     @classmethod
     def setUpClass(cls):
         cls.default_test_url = """https://raw.githubusercontent.com/hed-standard/hed-specification/HED-restructure/hedxml/HED7.1.1.xml"""
-        cls.hed_xml_file = os.path.join(os.path.dirname(os.path.abspath(__file__)), 'data/HED7.1.1.xml')
+        cls.hed_xml_file = os.path.join(os.path.dirname(os.path.abspath(__file__)), '../data/HED7.1.1.xml')
 
     def test_url_to_file(self):
         downloaded_file = file_util.url_to_file(self.default_test_url)
