+++ hedweb/deploy_hed/Dockerfile
@@ -19,7 +19,7 @@ COPY ./hedtools/hed/web /var/www/hedtools/hed/web/
 RUN mkdir -p /run/apache2 && \
 mkdir -p /var/log/hedtools && \
 chown -R apache:apache /var/log/hedtools && \
-mkdir -p /var/cache/hed_cache && \
-chown -R apache:apache /var/cache/hed_cache
+mkdir -p /var/cache/schema_cache && \
+chown -R apache:apache /var/cache/schema_cache
 ENTRYPOINT /usr/sbin/httpd -D FOREGROUND -f /etc/apache2/httpd.conf
 ENV HEDTOOLS_CONFIG_CLASS=config.ProductionConfig
